styled-components under the export component
