+++ client/src/components/Profile/ProfileSettings.js
@@ -2,7 +2,6 @@ import React from 'react'
 import TopHeader from '../../containers/TopHeaderContainer'
 import Footer from '../Footer/Footer'
 import styled from 'styled-components'
-
 import Loadable from 'react-loadable'
 import LoadingComponent from '../common/Loading'
